[Frontend] Moved AuthContext to contexts folder
